+++ src/main/java/wooteco/subway/ui/PathController.java
@@ -1,11 +1,13 @@
 package wooteco.subway.ui;
 
+import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.dto.PathResponse;
+import wooteco.subway.dto.path.PathRequest;
+import wooteco.subway.dto.path.PathResponse;
 import wooteco.subway.service.PathService;
 
 @RestController
@@ -19,8 +21,8 @@ public class PathController {
     }
 
     @GetMapping
-    public ResponseEntity<PathResponse> findPath(@RequestParam Long source, @RequestParam Long target,
-                                                 @RequestParam int age) {
-        return ResponseEntity.ok(pathService.findPath(source, target, age));
+    public ResponseEntity<PathResponse> findPath(@ModelAttribute @Valid PathRequest pathRequest) {
+        return ResponseEntity.ok(
+                pathService.findPath(pathRequest.getSource(), pathRequest.getTarget(), pathRequest.getAge()));
     }
 }
